tree: update

+++ wol/install.sh
@@ -1,4 +1,10 @@
 #!/usr/bin/env ash
+#
+# Copyright (C) 2023 AuxXxilium <https://github.com/AuxXxilium> and Ing <https://github.com/wjz304>
+#
+# This is free software, licensed under the MIT License.
+# See /LICENSE for more information.
+#
 
 if [ "${1}" = "jrExit" ]; then
   echo "Installing addon wol - ${1}"
@@ -10,29 +16,31 @@ elif [ "${1}" = "late" ]; then
   mkdir -p "/tmpRoot/usr/arc/addons/"
   cp -vf "${0}" "/tmpRoot/usr/arc/addons/"
   
-  cp -vf /usr/bin/ethtool /tmpRoot/usr/bin/ethtool
+  [ ! -f "/tmpRoot/usr/bin/ethtool" ] && cp -vf /usr/bin/ethtool /tmpRoot/usr/bin/ethtool
   cp -vf /usr/bin/wol.sh /tmpRoot/usr/bin/wol.sh
 
-  DEST="/tmpRoot/lib/systemd/system/ethtool.service"
-  echo "[Unit]"                                                               > ${DEST}
-  echo "Description=Arc force WoL on ethN"                                    >>${DEST}
-  echo                                                                        >>${DEST}
-  echo "[Service]"                                                            >>${DEST}
-  echo "Type=oneshot"                                                         >>${DEST}
-  echo "RemainAfterExit=yes"                                                  >>${DEST}
-  echo "ExecStart=/usr/bin/wol.sh"                                            >>${DEST}
-  echo                                                                        >>${DEST}
-  echo "[Install]"                                                            >>${DEST}
-  echo "WantedBy=multi-user.target"                                           >>${DEST}
+  mkdir -p "/tmpRoot/usr/lib/systemd/system"
+  DEST="/tmpRoot/usr/lib/systemd/system/wol.service"
+  echo "[Unit]"                                   > ${DEST}
+  echo "Description=ARPL force WoL on ethN"       >>${DEST}
+  echo "After=multi-user.target"                  >>${DEST}
+  echo                                            >>${DEST}
+  echo "[Service]"                                >>${DEST}
+  echo "Type=oneshot"                             >>${DEST}
+  echo "RemainAfterExit=yes"                      >>${DEST}
+  echo "ExecStart=/usr/bin/wol.sh"                >>${DEST}
+  echo                                            >>${DEST}
+  echo "[Install]"                                >>${DEST}
+  echo "WantedBy=multi-user.target"               >>${DEST}
 
-  mkdir -vp /tmpRoot/lib/systemd/system/multi-user.target.wants
-  ln -vsf /lib/systemd/system/ethtool.service /tmpRoot/lib/systemd/system/multi-user.target.wants/ethtool.service
+  mkdir -vp /tmpRoot/usr/lib/systemd/system/multi-user.target.wants
+  ln -vsf /usr/lib/systemd/system/wol.service /tmpRoot/usr/lib/systemd/system/multi-user.target.wants/wol.service
 elif [ "${1}" = "uninstall" ]; then
   echo "Installing addon wol - ${1}"
 
-  rm -f "/tmpRoot/lib/systemd/system/multi-user.target.wants/ethtool.service"
-  rm -f "/tmpRoot/lib/systemd/system/ethtool.service"
+  rm -f "/tmpRoot/usr/lib/systemd/system/multi-user.target.wants/wol.service"
+  rm -f "/tmpRoot/usr/lib/systemd/system/wol.service"
 
-  #rm -f /tmpRoot/usr/bin/ethtool
+  # rm -f /tmpRoot/usr/bin/ethtool
   rm -f /tmpRoot/usr/bin/wol.sh
 fi
